requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
torch
torchvision
transformers
tqdm
scikit-learn
pillow
fastapi
uvicorn
opencv-python
matplotlib
python-multipart